feat: footer, se hooks
